Address second round of review feedback
- Add assertUnreachable to DocumentViewer switch for exhaustive checking
- Harden address.name guard in BaseEmailViewer to handle null/undefined
- Add console.warn on TIFF-to-PDF transform failure
- Add @default JSDoc tags to all optional props across viewer API files
- Rename types.ts to <Name>Api.ts across all 7 viewers per CONTRIBUTING.md
- Document 30 new CSS variables in docs/CSSVariables.md
- Add new viewers to AGENTS.md, README.md, and sidebarsReactComponents.ts
