+++ packages/react-components/src/public/experimental/document-viewer.ts
@@ -21,5 +21,5 @@ export const DocumentViewer: typeof _DocumentViewer = withOsdkMetrics(
   _DocumentViewer,
   "DocumentViewer",
 );
-export type { DocumentViewerProps } from "../../document-viewer/types.js";
-export { ViewerType } from "../../document-viewer/types.js";
+export type { DocumentViewerProps } from "../../document-viewer/DocumentViewerApi.js";
+export { ViewerType } from "../../document-viewer/DocumentViewerApi.js";
